Rename app interface to wrapper
