+++ internal/wrapper/linter/linter.go
@@ -1,4 +1,4 @@
-package lintapp
+package linter
 
 import (
 	"context"
@@ -6,38 +6,38 @@ import (
 	"os"
 	"os/exec"
 
-	"github.com/anttiharju/vmatch/pkg/app"
 	"github.com/anttiharju/vmatch/pkg/exit"
 	"github.com/anttiharju/vmatch/pkg/exit/exitcode"
 	"github.com/anttiharju/vmatch/pkg/pathfinder"
 	"github.com/anttiharju/vmatch/pkg/versionfinder"
+	"github.com/anttiharju/vmatch/pkg/wrapper"
 )
 
-type LintApp struct {
+type WrappedLinter struct {
 	desiredVersion string
 	installPath    string
 }
 
-var _ app.Interface = (*LintApp)(nil)
+var _ wrapper.Interface = (*WrappedLinter)(nil)
 
-func NewApp(versionFileName string) *LintApp {
-	desiredVersion := versionfinder.GetVersion(pathfinder.GetWorkDir(), versionFileName)
+func NewWrapper() *WrappedLinter {
+	desiredVersion := versionfinder.GetVersion(pathfinder.GetWorkDir(), ".golangci-version")
 	installPath := pathfinder.GetInstallPath(desiredVersion)
 
-	return &LintApp{
+	return &WrappedLinter{
 		desiredVersion: desiredVersion,
 		installPath:    installPath,
 	}
 }
 
-func (l *LintApp) Run(ctx context.Context) int {
-	if l.noBinary() {
-		l.install(ctx)
+func (w *WrappedLinter) Run(ctx context.Context) int {
+	if w.noBinary() {
+		w.install(ctx)
 	}
 
 	args := os.Args[1:]
 	//nolint:gosec // I don't think a wrapper can avoid G204.
-	linter := exec.Command(l.getGolangCILintPath(), args...)
+	linter := exec.Command(w.getGolangCILintPath(), args...)
 	linterOutput, _ := linter.Output()
 
 	fmt.Print(string(linterOutput))
@@ -45,19 +45,19 @@ func (l *LintApp) Run(ctx context.Context) int {
 	return linter.ProcessState.ExitCode()
 }
 
-func (l *LintApp) noBinary() bool {
-	_, err := os.Stat(l.getGolangCILintPath())
+func (w *WrappedLinter) noBinary() bool {
+	_, err := os.Stat(w.getGolangCILintPath())
 
 	return os.IsNotExist(err)
 }
 
-func (l *LintApp) install(_ context.Context) {
+func (w *WrappedLinter) install(_ context.Context) {
 	//nolint:lll // Official binary install command:
 	// curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin v1.59.1
 	curl := "curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh"
 	pipe := " | "
 	sh := "sh -s -- -b "
-	command := curl + pipe + sh + l.installPath + " v" + l.desiredVersion
+	command := curl + pipe + sh + w.installPath + " v" + w.desiredVersion
 	cmd := exec.Command("sh", "-c", command)
 
 	err := cmd.Start()
@@ -71,6 +71,6 @@ func (l *LintApp) install(_ context.Context) {
 	}
 }
 
-func (l *LintApp) getGolangCILintPath() string {
-	return l.installPath + string(os.PathSeparator) + "golangci-lint"
+func (w *WrappedLinter) getGolangCILintPath() string {
+	return w.installPath + string(os.PathSeparator) + "golangci-lint"
 }
