--- pkg/app/app.go
@@ -1,9 +1,0 @@
-package app
-
-import "context"
-
-type Interface interface {
-	Run(ctx context.Context) int
-}
-
-type NewAppInterface func(versionFileName string) Interface
