+++ pkg/wrapper/wrapper.go
@@ -1,0 +1,13 @@
+package wrapper
+
+import "context"
+
+type RunInterface interface {
+	Run(ctx context.Context) int
+}
+
+type Interface interface {
+	RunInterface
+}
+
+type NewWrapper func() Interface
